requirements.txt
asttokens
beautifulsoup4
ensure
exceptiongroup
executing
fonttools
idna
ipykernel
ipython
joblib
matplotlib
matplotlib-inline
nltk
numpy
pandas
python-box
pytz
pywin32
PyYAML
pyzmq
regex
requests
scikit-learn
scipy
six
soupsieve
stack-data
threadpoolctl
tornado
tqdm
traitlets
typing_extensions
tzdata
urllib3
wcwidth
feedparser
torch
transformers
textblob